+++ packages/@expo/cli/src/run/ios/appleDevice/AppleDevice.ts
@@ -41,9 +41,13 @@ export interface ConnectedDevice {
 async function getConnectedDevicesUsingNativeToolsAsync(): Promise<ConnectedDevice[]> {
   return (
     (await devicectl.getConnectedAppleDevicesAsync())
-      // Filter out unpaired devices.
+      // Filter out unpaired and unavailable devices.
       // TODO: We could improve this logic in the future to attempt pairing if specified.
-      .filter((device) => device.connectionProperties.pairingState === 'paired')
+      .filter(
+        (device) =>
+          device.connectionProperties.pairingState === 'paired' &&
+          device.connectionProperties.tunnelState !== 'unavailable'
+      )
       .map((device) => {
         return {
           name: device.deviceProperties.name,