+++ packages/expo-modules-core/CHANGELOG.md
@@ -6,9 +6,15 @@
 
 ### 🎉 New features
 
+- Added `onKeyDown` and `onKeyLongPress` to `ReactActivityHandler` on Android. ([#28684](https://github.com/expo/expo/pull/28684) by [@lukmccall](https://github.com/lukmccall))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others
+
+## 1.12.9 — 2024-05-06
+
+_This version does not introduce any user-facing changes._
 
 ## 1.12.8 — 2024-05-03